update utility for docker images and more

+++ docker_run.sh
@@ -34,7 +34,13 @@ else
 fi
 # copy over trakman directory
 if find /app/server/trakman -mindepth 1 -maxdepth 1 | read; then
-  echo 'Trakman exists, skipping initial setup.'
+  echo 'Trakman exists. Attempting update...'
+  cd trakman || exit
+  node Update.js /app/server/trakmanbk/.hashes.json
+  if [ $? -gt 0 ]; then
+    exit
+  fi
+  cd ..
   rm -r trakmanbk
 else
   echo 'Setting up trakman...'
